refactor: move logger function from `app.py` to `utils/logger.py`
just some simply refactoring to clean up app.py
